add component for AWS VPC
