+++ libs/pages/clusters/src/lib/ui/page-clusters-create/step-features/step-features.spec.tsx
@@ -1,7 +1,7 @@
 import { wrapWithReactHookForm } from '__tests__/utils/wrap-with-react-hook-form'
 import { CloudProviderEnum } from 'qovery-typescript-axios'
 import { renderWithProviders, screen } from '@qovery/shared/util-tests'
-import StepFeatures, { type StepFeaturesProps, areSubnetsEmpty } from './step-features'
+import StepFeatures, { type StepFeaturesProps } from './step-features'
 
 const STATIC_IP = 'STATIC_IP'
 
@@ -73,15 +73,5 @@ describe('StepFeatures', () => {
 
     expect(button).not.toBeDisabled()
     expect(props.onSubmit).toHaveBeenCalled()
-  })
-
-  it('returns false if subnet has empty A, B, or C properties', () => {
-    const subnets = [{ A: '', B: '', C: '' }]
-    expect(areSubnetsEmpty(subnets)).toBe(true)
-  })
-
-  it('returns true if subnet has non-empty A, B, and C properties', () => {
-    const subnets = [{ A: 'subnet1A', B: 'subnet1B', C: '' }]
-    expect(areSubnetsEmpty(subnets)).toBe(false)
   })
 })
